[Reorg] Rename nifproxy .h/.cpp
