+++ src/model/nifproxymodel.cpp
@@ -30,7 +30,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 ***** END LICENCE BLOCK *****/
 
-#include "nifproxy.h"
+#include "nifproxymodel.h"
 
 #include "nifmodel.h"
