+++ src/model/nifproxymodel.h
